+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Removed mesos security from History Server and unified environment variable VAULT_HOSTS
 * Secret folder path configurable
 * Changed log format according to Stratio standards
+* Mesos Role no longer obtain mesos pricipal and mesos secret from vault in Spark jobs
 
 ## 2.2.0.4 (January 11, 2018)